+++ frappe/automation/doctype/auto_repeat/auto_repeat.py
@@ -8,7 +8,7 @@ from dateutil.relativedelta import relativedelta
 import frappe
 from frappe import _
 from frappe.automation.doctype.assignment_rule.assignment_rule import get_repeated
-from frappe.communications.doctype.communication.email import make
+from frappe.communications.email import make
 from frappe.contacts.doctype.contact.contact import (
 	get_contacts_linked_from,
 	get_contacts_linking_to,
